Move temp folder to core logic, move HiveServer2 tests into core classes
